Immediate conflicts resolved:
	rmgpy/molecule/adjlist.py
	rmgpy/qm/mopac.py
	rmgpy/reaction.py

But not yet checked that everything else works.

+++ Makefile
@@ -26,10 +26,10 @@ solver:
 
 ifneq ($(DASPK),)
 	@ echo "DASPK solver found. Compiling with DASPK and sensitivity analysis capability..."
-	@ echo "DEF DASPK = 1" > rmgpy/solver/settings.pxi 
+	@ (echo DEF DASPK = 1) > rmgpy/solver/settings.pxi 
 else ifneq ($(DASSL),)
 	@ echo "DASSL solver found. Compiling with DASSL.  Sensitivity analysis capabilities are off..."
-	@ echo "DEF DASPK = 0" > rmgpy/solver/settings.pxi
+	@ (echo DEF DASPK = 0) > rmgpy/solver/settings.pxi
 else
 	@ echo 'No PyDAS solvers found.  Please check if you have the latest version of PyDAS.'
 	@ python -c 'import pydas.dassl' 
@@ -39,11 +39,9 @@ endif
 cantherm:
 	python setup.py build_ext cantherm --build-lib . --build-temp build --pyrex-c-in-temp
 
-bin/symmetry:
-	mkdir -p bin
-	$(MAKE) -C external/symmetry install
-
-QM: bin/symmetry
+QM:
+	@ echo "Checking if you have symmetry..."
+	@ echo "symmetry -h"
 	@ echo "Checking you have rdkit..."
 	@ python -c 'import rdkit; print rdkit.__file__'
 	@ echo "Checking rdkit version..."
@@ -64,9 +62,7 @@ clean:
 	rm -rf build/
 	find . -name '*.so' -exec rm -f '{}' \;
 	find . -name '*.pyc' -exec rm -f '{}' \;
-	$(MAKE) -C external/symmetry clean
-	rm -f bin/symmetry
-
+	
 clean-solver:
 	rm -r build/pyrex/rmgpy/solver/
 	rm -r build/build/pyrex/rmgpy/solver/
@@ -79,63 +75,74 @@ decython:
 	find . -name *.pyc -exec rm -f '{}' \;
 
 test:
+ifeq ($(OS),Windows_NT)
+	nosetests --nocapture --nologcapture --all-modules --verbose --with-coverage --cover-inclusive --cover-package=rmgpy --cover-erase --cover-html --cover-html-dir=testing/coverage --exe rmgpy
+else
 	mkdir -p testing/coverage
 	rm -rf testing/coverage/*
 	nosetests --nocapture --nologcapture --all-modules --verbose --with-coverage --cover-inclusive --cover-package=rmgpy --cover-erase --cover-html --cover-html-dir=testing/coverage --exe rmgpy
-
+endif
 test-database:
 	nosetests -v -d testing/databaseTest.py	
 
 eg1: noQM
-	mkdir -p testing/minimal
-	rm -rf testing/minimal/*
-	cp examples/rmg/minimal/input.py testing/minimal/input.py
+	mkdir -p testing/eg1
+	rm -rf testing/eg1/*
+	cp examples/rmg/minimal/input.py testing/eg1/input.py
 	coverage erase
-	@ echo "Running minimal example with coverage tracking AND profiling"
-	coverage run rmg.py -p testing/minimal/input.py
+	@ echo "Running eg1: minimal (ethane pyrolysis) example with coverage tracking AND profiling"
+	coverage run rmg.py -p testing/eg1/input.py
 	coverage report
 	coverage html
 eg2: all
-	mkdir -p testing/hexadiene
-	rm -rf testing/hexadiene/*
-	cp examples/rmg/1,3-hexadiene/input.py testing/hexadiene/input.py
+	mkdir -p testing/eg2
+	rm -rf testing/eg2/*
+	cp examples/rmg/1,3-hexadiene/input.py testing/eg2/input.py
 	coverage erase
-	@ echo "Running 1,3-hexadiene example with coverage tracking AND profiling"
-	coverage run rmg.py -p testing/hexadiene/input.py
+	@ echo "Running eg2: 1,3-hexadiene example with coverage tracking AND profiling"
+	coverage run rmg.py -p testing/eg2/input.py
 	coverage report
 	coverage html
 
 eg3: all
-	mkdir -p testing/liquid_phase
-	rm -rf testing/liquid_phase/*
-	cp examples/rmg/liquid_phase/input.py testing/liquid_phase/input.py
+	mkdir -p testing/eg3
+	rm -rf testing/eg3/*
+	cp examples/rmg/liquid_phase/input.py testing/eg3/input.py
 	coverage erase
-	@ echo "Running liquid_phase example with coverage tracking AND profiling"
-	coverage run rmg.py -p testing/liquid_phase/input.py
+	@ echo "Running eg3: liquid_phase example with coverage tracking AND profiling"
+	coverage run rmg.py -p testing/eg3/input.py
 	coverage report
 	coverage html
 
 eg5: all
-	mkdir -p testing/heptane-eg5
-	rm -rf testing/heptane-eg5/*
-	cp examples/rmg/heptane-eg5/input.py testing/heptane-eg5/input.py
-	@ echo "Running heptane-eg5 example."
-	python rmg.py -q testing/heptane-eg5/input.py
+	mkdir -p testing/eg5
+	rm -rf testing/eg5/*
+	cp examples/rmg/heptane-eg5/input.py testing/eg5/input.py
+	@ echo "Running eg5: heptane example"
+	python rmg.py testing/eg5/input.py
 
 eg6: all
-	mkdir -p testing/ethane-oxidation
-	rm -rf testing/ethane-oxidation/*
-	cp examples/rmg/ethane-oxidation/input.py testing/ethane-oxidation/input.py
-	@ echo "Running ethane-oxidation example."
-	python rmg.py -q testing/ethane-oxidation/input.py
+	mkdir -p testing/eg6
+	rm -rf testing/eg6/*
+	cp examples/rmg/ethane-oxidation/input.py testing/eg6/input.py
+	@ echo "Running eg6: ethane-oxidation example"
+	python rmg.py testing/eg6/input.py
 
 eg7: all
-	mkdir -p testing/gri_mech_rxn_lib
-	rm -rf testing/gri_mech_rxn_lib/*
-	cp examples/rmg/gri_mech_rxn_lib/input.py testing/gri_mech_rxn_lib/input.py
-	@ echo "Running gri_mech_rxn_lib example."
-	python rmg.py testing/gri_mech_rxn_lib/input.py
+	mkdir -p testing/eg7
+	rm -rf testing/eg7/*
+	cp examples/rmg/gri_mech_rxn_lib/input.py testing/eg7/input.py
+	@ echo "Running eg7: gri_mech_rxn_lib example"
+	python rmg.py testing/eg7/input.py
 	
+scoop: noQM
+	mkdir -p testing/scoop
+	rm -rf testing/scoop/*
+	cp examples/rmg/minimal/input.py testing/scoop/input.py
+	coverage erase
+	@ echo "Running minimal example with SCOOP"
+	python -m scoop -n 2 rmg.py -v testing/scoop/input.py
+
 ######### 
 # Section for setting up MOPAC calculations on the Travis-CI.org server
 ifeq ($(TRAVIS),true)
@@ -161,9 +168,9 @@ eg4: all mopac_travis
 ifeq ($(SKIP_MOPAC),true)
 	@echo "Skipping eg4 (without failing) because can't run MOPAC"
 else
-	mkdir -p testing/thermoEstimator
-	rm -rf testing/thermoEstimator/*
-	cp examples/thermoEstimator/input.py testing/thermoEstimator/input.py
+	mkdir -p testing/eg4
+	rm -rf testing/eg4/*
+	cp examples/thermoEstimator/input.py testing/eg4/input.py
 	@ echo "Running thermo data estimator example. This tests QM."
-	python scripts/thermoEstimator.py testing/thermoEstimator/input.py
+	python scripts/thermoEstimator.py testing/eg4/input.py
 endif
